requirements.txt
streamlit
tensorflow>=2.11.0
numpy
pandas
matplotlib
scikit-learn
Pillow
